Create MyEventsComponent
